Helper to add label columns

Helper to add label columns

Usage

add_label_columns(
  results_table,
  columns_var_to_convert,
  dictionary_survey,
  dictionary_survey_name_column = "name",
  dictionary_choices,
  dictionary_choices_survey_column = "name_survey",
  dictionary_choices_choices_column = "name_choices"
)

Arguments

results_table

Result object with an analysis key.

columns_var_to_convert

set of columns to add labels.

dictionary_survey

Dictionary with name and label to use. Should be created with create_label_dictionary.

dictionary_survey_name_column

name column in the dictionary.

dictionary_choices

Dictionary with list_name, name and label to use. Should be created with create_label_dictionary.

dictionary_choices_survey_column

name column for the survey variable column in the dictionary.

dictionary_choices_choices_column

name column for the choices value column in the dictionary.

Value

Results table with label_* columns