Helper to check the results for the IPCTWG table

Helper to make sure the results contains all the IPCTWG information

Usage

check_ipctwg_results(
  .results,
  analysis_key = "analysis_key",
  fclc_matrix_var = "fclc_phase",
  fclc_matrix_values = c("Phase 1 - FCLC", "Phase 2 - FCLC", "Phase 3 - FCLC",
    "Phase 4 - FCLC", "Phase 5 - FCLC"),
  fc_matrix_var = "fc_phase",
  fc_matrix_values = c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC",
    "Phase 5 FC"),
  with_fclc = FALSE,
  fcs_cat_var = "fcs_cat",
  fcs_cat_values = c("Poor", "Borderline", "Acceptable"),
  rcsi_cat_var = "rcsi_cat",
  rcsi_cat_values = c("No to Low", "Medium", "High"),
  lcsi_cat_var = "lcsi_cat",
  lcsi_cat_values = c("None", "Stress", "Crisis", "Emergency"),
  hhs_cat_var = "hhs_cat_ipc",
  hhs_cat_values = c("None", "Little", "Moderate", "Severe", "Very Severe"),
  fcs_set = c("fs_fcs_cereals_grains_roots_tubers", "fs_fcs_beans_nuts", "fs_fcs_dairy",
    "fs_fcs_meat_fish_eggs", "fs_fcs_vegetables_leaves", "fs_fcs_fruit",
    "fs_fcs_oil_fat_butter", "fs_fcs_sugar", "fs_fcs_condiment"),
  rcsi_set = c("rCSILessQlty", "rCSIBorrow", "rCSIMealSize", "rCSIMealAdult",
    "rCSIMealNb"),
  lcsi_set,
  lcsi_value_set = c("yes", "no_had_no_need", "no_exhausted", "not_applicable"),
  other_variables = NULL,
  stat_col = "stat",
  proportion_name = "prop_select_one",
  mean_name = "mean"
)

Arguments

.results

results table with analysis key

analysis_key

String with the name of the analysis key. Default is "analysis_key"

fclc_matrix_var

String with the name of the food consumption livelihood matrix from FEWSNET. Default is "fclc_phase"

fclc_matrix_values

String with the options of the food consumption livelihood matrix Default is c("Phase 1 - FCLC", "Phase 2 - FCLC", "Phase 3 - FCLC", "Phase 4 - FCLC", "Phase 5 - FCLC")

fc_matrix_var

String with the name of the food consumption matrix from FEWSNET. Default is "fc_phase"

fc_matrix_values

String with the options of the food consumption matrix. Default is c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC", "Phase 5 FC")

with_fclc

TRUE or FALSE, whether to include the FCLC and FC values. Default is set to FALSE.

fcs_cat_var

The string with the name of the Food Consumption Score. Default is "fcs_cat"

fcs_cat_values

String with the options of the Food Consumption Score. Default is c("Poor", "Borderline", "Acceptable")

rcsi_cat_var

The string with the name of the reduced Coping Strategy Index. Default is "rcsi_cat"

rcsi_cat_values

String with the options of the reduced Coping Strategy Index. Default is c("No to Low", "Medium", "High")

lcsi_cat_var

The string with the name of the Livelihood Coping Strategy Index. Default is "lcsi_cat"

lcsi_cat_values

String with the options of the Livelihood Coping Strategy Index. Default is c("None", "Stress", "Crisis", "Emergency")

hhs_cat_var

The string with the name of the Household Hunger Scale. Default is "hhs_cat_ipc"

hhs_cat_values

String with the options of the Household Hunger Scale. Default is c("None", "Little", "Moderate", "Severe", "Very Severe")

fcs_set

String for the Food Consumption Score questions set. Default is c("fs_fcs_cereals_grains_roots_tubers", "fs_fcs_beans_nuts", "fs_fcs_dairy", "fs_fcs_meat_fish_eggs", "fs_fcs_vegetables_leaves", "fs_fcs_fruit", "fs_fcs_oil_fat_butter", "fs_fcs_sugar", "fs_fcs_condiment")

rcsi_set

String for the reduced Coping Strategy Index questions set. Default is c("No to Low", "Medium", "High")

lcsi_set

String for the Livelihood Coping Strategy Index questions set. There is no default

lcsi_value_set

String for the values of the Livelihood Coping Strategy Index questions set. Default is c("yes", "no_had_no_need", "no_exhausted", "not_applicable")

other_variables

String for the names of other variables to include. Default is NULL

stat_col

String for the name of the column with the mean or the proportion. Default is "stat"

proportion_name

String how a proportion is called in the analysis key. Default is "prop_select_one"

mean_name

String how a mean is called in the analysis key. Default is "mean"

Value

a list for 3 objects:

dictionary: list of all the variables and their values.

ipctw_key_table: key_table from the analysis_key.

value_checked_ipctw_key_table: list of key_table where all values are checked for each variables.

Examples

check_ipctwg_results(
  .results = presentresults_resultstable,
  fclc_matrix_var = "fcls_cat",
  fclc_matrix_values = c("phase_1", "phase_2", "phase_3", "phase_4", "phase_5"),
  fc_matrix_var = "fcm_cat",
  fc_matrix_values = c("phase_1", "phase_2", "phase_3", "phase_4", "phase_5"),
  with_fclc = TRUE,
  fcs_cat_values = c("low", "medium", "high"),
  rcsi_cat_values = c("low", "medium", "high"),
  lcsi_cat_var = "lcs_cat",
  lcsi_cat_values = c("none", "stress", "emergency", "crisis"),
  hhs_cat_var = "hhs_cat",
  hhs_cat_values = c("none", "slight", "moderate", "severe", "very_severe"),
  lcsi_set = c(
    "liv_stress_lcsi_1",
    "liv_stress_lcsi_2",
    "liv_stress_lcsi_3",
    "liv_stress_lcsi_4",
    "liv_crisis_lcsi_1",
    "liv_crisis_lcsi_2",
    "liv_crisis_lcsi_3",
    "liv_emerg_lcsi_1",
    "liv_emerg_lcsi_2",
    "liv_emerg_lcsi_3"
  )
)
#> $dictionary
#> $dictionary$fclc_matrix
#> $dictionary$fclc_matrix$var_name
#> [1] "fcls_cat"
#> 
#> $dictionary$fclc_matrix$values_set
#> [1] "phase_1" "phase_2" "phase_3" "phase_4" "phase_5"
#> 
#> $dictionary$fclc_matrix$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$fc_matrix
#> $dictionary$fc_matrix$var_name
#> [1] "fcm_cat"
#> 
#> $dictionary$fc_matrix$values_set
#> [1] "phase_1" "phase_2" "phase_3" "phase_4" "phase_5"
#> 
#> $dictionary$fc_matrix$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$fcs
#> $dictionary$fcs$var_name
#> [1] "fcs_cat"
#> 
#> $dictionary$fcs$values_set
#> [1] "low"    "medium" "high"  
#> 
#> $dictionary$fcs$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$rcsi
#> $dictionary$rcsi$var_name
#> [1] "rcsi_cat"
#> 
#> $dictionary$rcsi$values_set
#> [1] "low"    "medium" "high"  
#> 
#> $dictionary$rcsi$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$lcsi
#> $dictionary$lcsi$var_name
#> [1] "lcs_cat"
#> 
#> $dictionary$lcsi$values_set
#> [1] "none"      "stress"    "emergency" "crisis"   
#> 
#> $dictionary$lcsi$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$hhs
#> $dictionary$hhs$var_name
#> [1] "hhs_cat"
#> 
#> $dictionary$hhs$values_set
#> [1] "none"        "slight"      "moderate"    "severe"      "very_severe"
#> 
#> $dictionary$hhs$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$fcs_set_mean
#> $dictionary$fcs_set_mean$var_name
#> [1] "fs_fcs_cereals_grains_roots_tubers" "fs_fcs_beans_nuts"                 
#> [3] "fs_fcs_dairy"                       "fs_fcs_meat_fish_eggs"             
#> [5] "fs_fcs_vegetables_leaves"           "fs_fcs_fruit"                      
#> [7] "fs_fcs_oil_fat_butter"              "fs_fcs_sugar"                      
#> [9] "fs_fcs_condiment"                  
#> 
#> $dictionary$fcs_set_mean$values_set
#> [1] NA
#> 
#> $dictionary$fcs_set_mean$analysis_type
#> [1] "mean"
#> 
#> 
#> $dictionary$rcsi_set_mean
#> $dictionary$rcsi_set_mean$var_name
#> [1] "rCSILessQlty"  "rCSIBorrow"    "rCSIMealSize"  "rCSIMealAdult"
#> [5] "rCSIMealNb"   
#> 
#> $dictionary$rcsi_set_mean$values_set
#> [1] NA
#> 
#> $dictionary$rcsi_set_mean$analysis_type
#> [1] "mean"
#> 
#> 
#> $dictionary$rcsi_set_prop
#> $dictionary$rcsi_set_prop$var_name
#> [1] "rCSILessQlty"  "rCSIBorrow"    "rCSIMealSize"  "rCSIMealAdult"
#> [5] "rCSIMealNb"   
#> 
#> $dictionary$rcsi_set_prop$values_set
#> [1] 0 1 2 3 4 5 6 7
#> 
#> $dictionary$rcsi_set_prop$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> $dictionary$lcsi_set_prop
#> $dictionary$lcsi_set_prop$var_name
#>  [1] "liv_stress_lcsi_1" "liv_stress_lcsi_2" "liv_stress_lcsi_3"
#>  [4] "liv_stress_lcsi_4" "liv_crisis_lcsi_1" "liv_crisis_lcsi_2"
#>  [7] "liv_crisis_lcsi_3" "liv_emerg_lcsi_1"  "liv_emerg_lcsi_2" 
#> [10] "liv_emerg_lcsi_3" 
#> 
#> $dictionary$lcsi_set_prop$values_set
#> [1] "yes"            "no_had_no_need" "no_exhausted"   "not_applicable"
#> 
#> $dictionary$lcsi_set_prop$analysis_type
#> [1] "prop_select_one"
#> 
#> 
#> 
#> $ipctw_key_table
#> # A tibble: 893 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#>  2 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#>  3 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#>  4 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#>  5 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#>  6 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#>  7 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#>  8 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#>  9 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#> 10 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#> # ℹ 883 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table
#> $value_checked_ipctw_key_table$fclc_matrix
#> # A tibble: 30 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… fcls_cat     phase_1            location…
#>  2 prop_select_one @/@ … prop_select_… fcls_cat     phase_2            location…
#>  3 prop_select_one @/@ … prop_select_… fcls_cat     phase_3            location…
#>  4 prop_select_one @/@ … prop_select_… fcls_cat     phase_4            location…
#>  5 prop_select_one @/@ … prop_select_… fcls_cat     phase_5            location…
#>  6 prop_select_one @/@ … prop_select_… fcls_cat     phase_1            location…
#>  7 prop_select_one @/@ … prop_select_… fcls_cat     phase_2            location…
#>  8 prop_select_one @/@ … prop_select_… fcls_cat     phase_3            location…
#>  9 prop_select_one @/@ … prop_select_… fcls_cat     phase_4            location…
#> 10 prop_select_one @/@ … prop_select_… fcls_cat     phase_5            location…
#> # ℹ 20 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$fc_matrix
#> # A tibble: 29 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… fcm_cat      phase_1            location…
#>  2 prop_select_one @/@ … prop_select_… fcm_cat      phase_2            location…
#>  3 prop_select_one @/@ … prop_select_… fcm_cat      phase_3            location…
#>  4 prop_select_one @/@ … prop_select_… fcm_cat      phase_4            location…
#>  5 prop_select_one @/@ … prop_select_… fcm_cat      phase_5            location…
#>  6 prop_select_one @/@ … prop_select_… fcm_cat      phase_1            location…
#>  7 prop_select_one @/@ … prop_select_… fcm_cat      phase_2            location…
#>  8 prop_select_one @/@ … prop_select_… fcm_cat      phase_3            location…
#>  9 prop_select_one @/@ … prop_select_… fcm_cat      phase_4            location…
#> 10 prop_select_one @/@ … prop_select_… fcm_cat      phase_5            location…
#> # ℹ 19 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$fcs
#> # A tibble: 18 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#>  2 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#>  3 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#>  4 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#>  5 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#>  6 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#>  7 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#>  8 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#>  9 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#> 10 prop_select_one @/@ … prop_select_… fcs_cat      low                location…
#> 11 prop_select_one @/@ … prop_select_… fcs_cat      medium             location…
#> 12 prop_select_one @/@ … prop_select_… fcs_cat      high               location…
#> 13 prop_select_one @/@ … prop_select_… fcs_cat      low                location 
#> 14 prop_select_one @/@ … prop_select_… fcs_cat      medium             location 
#> 15 prop_select_one @/@ … prop_select_… fcs_cat      high               location 
#> 16 prop_select_one @/@ … prop_select_… fcs_cat      low                location 
#> 17 prop_select_one @/@ … prop_select_… fcs_cat      medium             location 
#> 18 prop_select_one @/@ … prop_select_… fcs_cat      high               location 
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$rcsi
#> # A tibble: 18 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… rcsi_cat     low                location…
#>  2 prop_select_one @/@ … prop_select_… rcsi_cat     medium             location…
#>  3 prop_select_one @/@ … prop_select_… rcsi_cat     high               location…
#>  4 prop_select_one @/@ … prop_select_… rcsi_cat     low                location…
#>  5 prop_select_one @/@ … prop_select_… rcsi_cat     medium             location…
#>  6 prop_select_one @/@ … prop_select_… rcsi_cat     high               location…
#>  7 prop_select_one @/@ … prop_select_… rcsi_cat     low                location…
#>  8 prop_select_one @/@ … prop_select_… rcsi_cat     medium             location…
#>  9 prop_select_one @/@ … prop_select_… rcsi_cat     high               location…
#> 10 prop_select_one @/@ … prop_select_… rcsi_cat     low                location…
#> 11 prop_select_one @/@ … prop_select_… rcsi_cat     medium             location…
#> 12 prop_select_one @/@ … prop_select_… rcsi_cat     high               location…
#> 13 prop_select_one @/@ … prop_select_… rcsi_cat     low                location 
#> 14 prop_select_one @/@ … prop_select_… rcsi_cat     medium             location 
#> 15 prop_select_one @/@ … prop_select_… rcsi_cat     high               location 
#> 16 prop_select_one @/@ … prop_select_… rcsi_cat     low                location 
#> 17 prop_select_one @/@ … prop_select_… rcsi_cat     medium             location 
#> 18 prop_select_one @/@ … prop_select_… rcsi_cat     high               location 
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$lcsi
#> # A tibble: 24 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… lcs_cat      none               location…
#>  2 prop_select_one @/@ … prop_select_… lcs_cat      stress             location…
#>  3 prop_select_one @/@ … prop_select_… lcs_cat      emergency          location…
#>  4 prop_select_one @/@ … prop_select_… lcs_cat      crisis             location…
#>  5 prop_select_one @/@ … prop_select_… lcs_cat      none               location…
#>  6 prop_select_one @/@ … prop_select_… lcs_cat      stress             location…
#>  7 prop_select_one @/@ … prop_select_… lcs_cat      emergency          location…
#>  8 prop_select_one @/@ … prop_select_… lcs_cat      crisis             location…
#>  9 prop_select_one @/@ … prop_select_… lcs_cat      none               location…
#> 10 prop_select_one @/@ … prop_select_… lcs_cat      stress             location…
#> # ℹ 14 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$hhs
#> # A tibble: 30 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… hhs_cat      none               location…
#>  2 prop_select_one @/@ … prop_select_… hhs_cat      slight             location…
#>  3 prop_select_one @/@ … prop_select_… hhs_cat      moderate           location…
#>  4 prop_select_one @/@ … prop_select_… hhs_cat      severe             location…
#>  5 prop_select_one @/@ … prop_select_… hhs_cat      very_severe        location…
#>  6 prop_select_one @/@ … prop_select_… hhs_cat      none               location…
#>  7 prop_select_one @/@ … prop_select_… hhs_cat      slight             location…
#>  8 prop_select_one @/@ … prop_select_… hhs_cat      moderate           location…
#>  9 prop_select_one @/@ … prop_select_… hhs_cat      severe             location…
#> 10 prop_select_one @/@ … prop_select_… hhs_cat      very_severe        location…
#> # ℹ 20 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$fcs_set_mean
#> # A tibble: 54 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 mean @/@ fs_fcs_cere… mean          fs_fcs_cere… NA                 location…
#>  2 mean @/@ fs_fcs_cere… mean          fs_fcs_cere… NA                 location…
#>  3 mean @/@ fs_fcs_cere… mean          fs_fcs_cere… NA                 location…
#>  4 mean @/@ fs_fcs_cere… mean          fs_fcs_cere… NA                 location…
#>  5 mean @/@ fs_fcs_bean… mean          fs_fcs_bean… NA                 location…
#>  6 mean @/@ fs_fcs_bean… mean          fs_fcs_bean… NA                 location…
#>  7 mean @/@ fs_fcs_bean… mean          fs_fcs_bean… NA                 location…
#>  8 mean @/@ fs_fcs_bean… mean          fs_fcs_bean… NA                 location…
#>  9 mean @/@ fs_fcs_dair… mean          fs_fcs_dairy NA                 location…
#> 10 mean @/@ fs_fcs_dair… mean          fs_fcs_dairy NA                 location…
#> # ℹ 44 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$rcsi_set_mean
#> # A tibble: 30 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 mean @/@ rCSILessQlt… mean          rCSILessQlty NA                 location…
#>  2 mean @/@ rCSILessQlt… mean          rCSILessQlty NA                 location…
#>  3 mean @/@ rCSILessQlt… mean          rCSILessQlty NA                 location…
#>  4 mean @/@ rCSILessQlt… mean          rCSILessQlty NA                 location…
#>  5 mean @/@ rCSIBorrow … mean          rCSIBorrow   NA                 location…
#>  6 mean @/@ rCSIBorrow … mean          rCSIBorrow   NA                 location…
#>  7 mean @/@ rCSIBorrow … mean          rCSIBorrow   NA                 location…
#>  8 mean @/@ rCSIBorrow … mean          rCSIBorrow   NA                 location…
#>  9 mean @/@ rCSIMealSiz… mean          rCSIMealSize NA                 location…
#> 10 mean @/@ rCSIMealSiz… mean          rCSIMealSize NA                 location…
#> # ℹ 20 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$rcsi_set_prop
#> # A tibble: 230 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… rCSILessQlty 0                  location…
#>  2 prop_select_one @/@ … prop_select_… rCSILessQlty 1                  location…
#>  3 prop_select_one @/@ … prop_select_… rCSILessQlty 2                  location…
#>  4 prop_select_one @/@ … prop_select_… rCSILessQlty 3                  location…
#>  5 prop_select_one @/@ … prop_select_… rCSILessQlty 4                  location…
#>  6 prop_select_one @/@ … prop_select_… rCSILessQlty 5                  location…
#>  7 prop_select_one @/@ … prop_select_… rCSILessQlty 6                  location…
#>  8 prop_select_one @/@ … prop_select_… rCSILessQlty 7                  location…
#>  9 prop_select_one @/@ … prop_select_… rCSILessQlty 0                  location…
#> 10 prop_select_one @/@ … prop_select_… rCSILessQlty 1                  location…
#> # ℹ 220 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#> $value_checked_ipctw_key_table$lcsi_set_prop
#> # A tibble: 240 × 8
#>    analysis_key          analysis_type analysis_var analysis_var_value group_var
#>    <chr>                 <chr>         <chr>        <chr>              <chr>    
#>  1 prop_select_one @/@ … prop_select_… liv_stress_… no_exhausted       location…
#>  2 prop_select_one @/@ … prop_select_… liv_stress_… no_had_no_need     location…
#>  3 prop_select_one @/@ … prop_select_… liv_stress_… not_applicable     location…
#>  4 prop_select_one @/@ … prop_select_… liv_stress_… yes                location…
#>  5 prop_select_one @/@ … prop_select_… liv_stress_… no_exhausted       location…
#>  6 prop_select_one @/@ … prop_select_… liv_stress_… no_had_no_need     location…
#>  7 prop_select_one @/@ … prop_select_… liv_stress_… not_applicable     location…
#>  8 prop_select_one @/@ … prop_select_… liv_stress_… yes                location…
#>  9 prop_select_one @/@ … prop_select_… liv_stress_… no_exhausted       location…
#> 10 prop_select_one @/@ … prop_select_… liv_stress_… no_had_no_need     location…
#> # ℹ 230 more rows
#> # ℹ 3 more variables: group_var_value <chr>, nb_analysis_var <dbl>,
#> #   nb_group_var <dbl>
#> 
#>