Write a table variable by group into Excel

Write a table variable by group into Excel

Usage

create_xlsx_variable_x_group(
  table_group_x_variable,
  file_path = NULL,
  table_name = "variable_x_group_table",
  value_columns = c("stat", "stat_low", "stat_upp"),
  total_columns = NULL,
  readme_sheet_name = "readme",
  table_sheet_name = "variable_x_group_table",
  overwrite = FALSE
)

Arguments

table_group_x_variable

a table create by create_table_variable_x_group

file_path

File names and path. Default is NULL which will return a workbook instead of an excel file.

table_name

string with the name of table to write. It will only be used if it is part of a list. Default is "variable_x_group_table".

value_columns

string containing the names of the columns with the stats to export

total_columns

string containing the names of the columns with the totals (n, n_total, n_weighted, n_total_weighted, etc.) to export

readme_sheet_name

string with the name of the sheet to write the read me page, default is "readme"

table_sheet_name

string with the name of the sheet to write the table, default is "variable_x_group_table"

overwrite

Default is FALSE, it will overwrite the file if set to TRUE.

Value

An excel file formatted.

Examples

if (FALSE) { # \dontrun{
presentresults_resultstable %>%
  create_table_variable_x_group() %>%
  create_xlsx_variable_x_group(file_path = "mytable.xlsx")
} # }