presentresults: Helpers to present results into table or graphs

The package provides wrappers to turn results from a long format into wide or graphics. For results in a long format, it will work with an analysis key which have the following format analysis_type @/@ dependent_variable

See also

Author

Maintainer: Yann Say yann.say@impact-initiatives.org (ORCID)

Authors:

Mehedi Khan mh.khan@reach-initiative.org (ORCID)